+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 ### Added
 - Invoice resource to load your account with dynamic QR Codes
 - Deposit resource to receive transfers passively
+- DictKey resource to get PIX key's parameters
 
 ## [2.1.0] - 2020-10-28
 ### Added